ui: Move some functions into new timeline::util module
