+++ crates/matrix-sdk-ui/src/timeline/inner/state.rs
@@ -41,8 +41,8 @@ use crate::{
         event_item::EventItemIdentifier,
         item::timeline_item,
         reactions::{ReactionToggleResult, Reactions},
-        rfind_event_item, timestamp_to_date,
         traits::RoomDataProvider,
+        util::{rfind_event_item, timestamp_to_date},
         AnnotationKey, Error as TimelineError, Profile, ReactionSenderData, TimelineItem,
         TimelineItemKind, VirtualTimelineItem,
     },
